+++ benchmarks/bench_queries_pmq.cpp
@@ -80,7 +80,7 @@ template <typename container_t>
 void run_bench(container_t& container, std::vector<elttype>& input_vec, const int batch_size, const int n_exp) {
 
    //create container
-   container.create(input_vec.size());
+   container.create((uint32_t)input_vec.size());
 
    // =====================================
    // Populates container and index
